refactor: useEmailInputForm 훅을 사용하여 state와 handler를 기존 컴포넌트에서 분리
